Windows and screens no longer behave cyclically
In default DWM, windows and screens have cyclic behavior, which means
that when they reach the bottom of the stack (or the last monitor), they
go back to the beginning. This has been disabled by default, although
options have been added to config.h so that it can be changed.

Some unnecessary functions like togglebar and incnmaster have been
removed from dwm.c