Mouse fixes and better window movement between monitors
Now mouse moves from one monitor to another when the user moves a window
to it. This also solves many crashes when moving the mouse between
monitors (my fault hehe).

Some new behaviour for windows was also added: now when moving a window
from one monitor to another, the window is put in the master position or
in the bottom of the stack depending on where the monitor is placed with
respect to the first one.